requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
google-genai>=0.8.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0
joblib==1.3.2
gunicorn==22.0.0
httpx==0.27.0